# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Quin Course",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/blogs/all": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get list blogs",
                "description": "Returns list blogs",
                "operationId": "getblogs",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        },
        "/api/blogs/{slug}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "description": "Returns blog information",
                "operationId": "Get blog detail",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Blog slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        },
        "/api/blogs/my_blogs": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "description": "Returns your list blogs",
                "operationId": "myblogs",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/blogs/{blog_id}/same_author": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "description": "Returns your list blogs",
                "operationId": "bogsameauthor",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "blog_id",
                        "in": "path",
                        "description": "bog_id og thif current blog",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/blogs/tag/{tag_slug}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "description": "Returns  list blogs by tag slug",
                "operationId": "blog_by_tag_slug",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "tag_slug",
                        "in": "path",
                        "description": "bog_id og thif current blog",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/blogs/create": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Create a new blog",
                "description": "Returns new new blog information",
                "operationId": "create_blog",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the new blog",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Content of the new blog",
                                        "type": "string"
                                    },
                                    "tag_id": {
                                        "description": "Tag ID of the new blog",
                                        "type": "integer"
                                    },
                                    "thumbnail_url": {
                                        "description": "Thumbnail blog ",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/blogs/update/{id}": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Update a new blog",
                "description": "Returns  blog information",
                "operationId": "update_blog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the blog to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the new blog",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Content of the new blog",
                                        "type": "string"
                                    },
                                    "tag_id": {
                                        "description": "Tag ID of the new blog",
                                        "type": "integer"
                                    },
                                    "thumbnail_url": {
                                        "description": "Thumbnail blog ",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/blogs/{blog_id}/delete_tag_blog/{tag_id}": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Delete tag blog",
                "description": "Returns  blog information",
                "operationId": "delete_tag_blog",
                "parameters": [
                    {
                        "name": "blog_id",
                        "in": "path",
                        "description": "ID of the blog ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag_id",
                        "in": "path",
                        "description": "ID of the tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/blogs/delete_blog/{blog_id}": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Delete  blog",
                "description": "Returns  blog information",
                "operationId": "delete_blog",
                "parameters": [
                    {
                        "name": "blog_id",
                        "in": "path",
                        "description": "ID of the blog ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Token (Bearer <token>)"
            }
        }
    }
}